Fix: Исправление конфликта обработчиков на мониторе Синего Щита.
